add folder for 3 view / separate test by type data / computed / methods / add test on undo function
