+++ tests/unit/Thead/Computed.spec.js
@@ -1,0 +1,36 @@
+import { mount } from '@vue/test-utils';
+import Thead from '@/components/Thead.vue';
+
+// data
+import exempleData from '@/data';
+
+let wrapper;
+
+beforeEach(() => {
+  const { headers } = exempleData;
+  const { tbodyIndex } = exempleData;
+  const { submenuThead } = exempleData;
+  const { disableSortThead } = exempleData;
+  const { sortHeader } = exempleData;
+  const submenuStatusThead = true;
+
+  wrapper = mount(Thead, {
+    propsData: {
+      submenuStatusThead,
+      submenuThead,
+      headers,
+      disableSortThead,
+      sortHeader,
+      tbodyIndex,
+    },
+  });
+  return wrapper;
+});
+
+describe('VueTable', () => {
+  describe('Render component with props', () => {
+    test('Vue Instance', () => {
+      expect(wrapper.isVueInstance()).toBeTruthy();
+    });
+  });
+});
